Persistent Vault, database pending.
This PR implements a persistent item and currency vault where you can insert coins and items, and be able to retrieve them in subsequent rounds. Currently, this persistence is handled by writing to a JSON backup file when the round ends, and loading from that same backup during the world loading phase.